Refactor service state management into dedicated components
