+++ Error Log/Build Errors.csv
@@ -44,3 +44,4 @@ XDG0008,DesktopApplicationTemplate.UI/FTPServiceView.xaml,2
 System.ArgumentException,DesktopApplicationTemplate.UI/CsvServiceCollectionExtensions.cs,1
 System.Windows.Markup.XamlParseException,DesktopApplicationTemplate.UI/MainWindow.xaml,1
 System.Windows.Markup.XamlParseException,DesktopApplicationTemplate.UI/TcpServiceMessagesView.xaml,1
+DOTNET_NOT_FOUND,DesktopApplicationTemplate/dotnet,1
