+++ Error Log/Code Errors.csv
@@ -4,3 +4,4 @@ XDG0008,60,2025-10-09T18:06:21
 System.NullReferenceException,1,2025-10-09T18:15:23
 System.ArgumentException,1,2025-10-09T18:26:39
 System.Windows.Markup.XamlParseException,2,2025-10-09T18:41:01
+DOTNET_NOT_FOUND,1,2025-10-10T17:51:59
