+++ Error Log/Detailed Errors.csv
@@ -54,3 +54,4 @@ System.NullReferenceException,DesktopApplicationTemplate.UI/App.xaml.cs,Object r
 System.ArgumentException,DesktopApplicationTemplate.UI/CsvServiceCollectionExtensions.cs,"Implementation type cannot be 'DesktopApplicationTemplate.Services.ServiceUiRegistration`2[DesktopApplicationTemplate.UI.ViewModels.ServiceListModel,System.Windows.Controls.Page]' because it is indistinguishable from other services registered for 'DesktopApplicationTemplate.Services.ServiceUiRegistration`2[DesktopApplicationTemplate.UI.ViewModels.ServiceListModel,System.Windows.Controls.Page]'.",1,2025-10-09T18:26:39
 System.Windows.Markup.XamlParseException,DesktopApplicationTemplate.UI/MainWindow.xaml,Add value to collection of type 'System.Windows.Controls.UIElementCollection' threw an exception. Line number '195' and line position '22'.,1,2025-10-09T18:34:29
 System.Windows.Markup.XamlParseException,DesktopApplicationTemplate.UI/TcpServiceMessagesView.xaml,Add value to collection of type 'System.Windows.Controls.UIElementCollection' threw an exception. Line number '60' and line position '18'.,1,2025-10-09T18:41:01
+DOTNET_NOT_FOUND,DesktopApplicationTemplate/dotnet,'dotnet' CLI is not installed; command 'dotnet --version' failed.,1,2025-10-10T17:51:59
